Add support for browsers
